reorganize files and folders
