+++ src/components/business_profile/readme.md
@@ -1,0 +1,1 @@
+This folder will contain all the component on business page.
