+++ src/components/login.js
@@ -21,7 +21,7 @@ class Login extends React.Component {
 
   render() {
     if (isLoggedIn()) {
-      navigate(`/app/profile`)
+      navigate(`/app/user-profile`)
     }
 
     return (
@@ -31,7 +31,7 @@ class Login extends React.Component {
           method="post"
           onSubmit={event => {
             this.handleSubmit(event)
-            navigate(`/app/profile`)
+            navigate(`/app/user-profile`)
           }}
         >
           <label>
