+++ src/components/user_profile/readme.md
@@ -1,0 +1,1 @@
+This folder will contain all the components for the user profile
